requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
gevent==24.2.1
greenlet==3.0.3
idna==3.7
isodate==0.6.1
logzero==1.7.0
mysql-connector==2.2.9
mysql-connector-python==9.0.0
numpy==1.24.4
pandas==2.0.3
pycrypto==2.6.1
pyotp==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
rdflib==7.0.0
rdflib-jsonld==0.6.2
requests==2.32.3
simplejson==3.19.3
six==1.16.0
SmartAPI==1.1.0
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
zope.event==5.0
zope.interface==7.0.1
